Umverlegung der Logik aus MonitorMC nach PlayerpathCommand
